update postcss custom var module to newer postcss version and esm
fixes Update postcss-custom-properties from v12 to v13 #442
