+++ packages/uui-css/package.json
@@ -37,7 +37,7 @@
     "build:cssprops": "postcss lib/custom-properties.css --base lib/ --dir dist && npm run custom-properties-cache",
     "clean": "tsc --build --clean && rimraf -g dist lib/*.js lib/**/*.js custom-elements.json custom-properties.*",
     "analyze": "web-component-analyzer **/*.element.ts --outFile custom-elements.json",
-    "custom-properties-cache": "node ./scripts/update-custom-properties-cache.mjs"
+    "custom-properties-cache": "node ./scripts/update-custom-properties-cache.js"
   },
   "publishConfig": {
     "access": "public"
