+++ packages/uui-css/scripts/update-custom-properties-cache.js
@@ -1,0 +1,6 @@
+import { CacheCustomProperties } from './cache-custom-properties.js';
+import path from 'path';
+
+const mainStylePath = path.resolve('./dist/custom-properties.css');
+
+CacheCustomProperties(mainStylePath);
