--- packages/uui-css/scripts/update-custom-properties-cache.mjs
@@ -1,6 +1,0 @@
-import { CacheCustomProperties } from './cache-custom-properties.mjs';
-import path from 'path';
-
-const mainStylePath = path.resolve('./dist/custom-properties.css');
-
-CacheCustomProperties(mainStylePath, true);
